refactor(chat): split workspace browser panel
Summary:
- split the Project AI workspace browser into sidebar, detail,
  helper, and state modules to satisfy the frontend file budget
- adjust the local workspace browser shell runner to feed scripts
  over stdin so the backend lint gate stays green
- record the new workspace browser read endpoints as intentional
  CLI/OpenAPI gaps until dedicated commands exist

Rationale:
- keep the new browser maintainable without one oversized Svelte
  component while preserving the shipped ASE-138 behavior
- keep repo-aligned validation green after the new endpoints landed
  by fixing the parity and gosec follow-up surfaced by CI

Tests:
- PLAYWRIGHT_WEB_PORT=4273 corepack pnpm --dir web run ci
- OPENASE_GO_TEST_PROGRESS_MODE=json OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest env OPENASE_BACKEND_TEST_GROUP_SIZE=8 make check
- PATH=$HOME/.local/go1.26.1/bin:$PATH go test ./internal/chat ./internal/httpapi ./internal/cli
- make build
- make LINT_BASE_REV=1258c9705f6b4514c06d6ea00be13927118d5a97 lint
- make lint-depguard
- make lint-architecture
