+++ internal/cli/api_parity_test.go
@@ -23,6 +23,10 @@ var intentionalCLIOpenAPIGaps = map[string]string{
 	contractKey("POST", "/api/v1/admin/auth/oidc-draft/test"):                                                      "instance OIDC connectivity testing has no first-class CLI yet",
 	contractKey("POST", "/api/v1/admin/auth/oidc-enable"):                                                          "instance OIDC activation is intentionally guided through the admin UI today",
 	contractKey("POST", "/api/v1/admin/auth/disable"):                                                              "instance auth rollback is intentionally guided through the admin UI today",
+	contractKey("GET", "/api/v1/chat/conversations/{conversationId}/workspace"):                                    "project conversation workspace inspection has no first-class CLI yet",
+	contractKey("GET", "/api/v1/chat/conversations/{conversationId}/workspace/file"):                               "project conversation workspace file preview has no first-class CLI yet",
+	contractKey("GET", "/api/v1/chat/conversations/{conversationId}/workspace/file-patch"):                         "project conversation workspace file diff has no first-class CLI yet",
+	contractKey("GET", "/api/v1/chat/conversations/{conversationId}/workspace/tree"):                               "project conversation workspace tree browsing has no first-class CLI yet",
 	contractKey("GET", "/api/v1/chat/projects/{projectId}/conversations/stream"):                                   "project-scoped conversation stream has no first-class CLI yet",
 	contractKey("GET", "/api/v1/instance/users"):                                                                   "instance user directory listing has no first-class CLI yet",
 	contractKey("GET", "/api/v1/instance/users/{userId}"):                                                          "instance user directory detail has no first-class CLI yet",
